backend/requirements.txt
# Python
Django==5.0.3
gunicorn
Pillow

# Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ

# Base de datos
psycopg2-binary